add more code for api routes. Add readme skeleton.
