+++ README.md
@@ -1,3 +1,78 @@
 # language-learn
-A web app to learn other languages!
-Hello!
+
+## Frontend Vue.js Application
+
+### Setting up the development environment
+
+### Running the application
+
+### Running Tests
+
+### Deploying to production
+
+### Project Structure
+```
+├── README.md
+├── index.html
+├── package-lock.json
+├── package.json
+├── public
+│   └── favicon.ico
+└── src
+    ├── App.vue
+    ├── assets
+    │   ├── base.css
+    │   ├── icons
+    │   │   └── logo.svg
+    │   └── main.css
+    ├── components
+    │   ├── Footer.vue
+    │   └── Navbar.vue
+    ├── main.js
+    ├── router
+    │   └── index.js
+    └── views
+        ├── Home.vue
+        ├── Login.vue
+        ├── MultipleChoiceQuiz.vue
+        └── Register.vue
+```
+
+## Backend Flask API
+
+### Setting up the development environment
+
+### Running the application
+
+### Running Tests
+
+### Deploying to production
+
+### Project Structure
+
+```
+├── app
+│   ├── __init__.py
+│   ├── models
+│   │   ├── __init__.py
+│   │   └── user.py
+│   ├── routes
+│   │   ├── __init__.py
+│   │   ├── auth.py
+│   │   ├── quiz.py
+│   │   └── user.py
+│   ├── services
+│   │   ├── __init__.py
+│   │   ├── auth_service.py
+│   │   ├── quiz_service.py
+│   │   └── user_service.py
+│   └── utils
+│       └── __init__.py
+├── app.db
+├── config.py
+├── requirements.txt
+└── sample_data.json
+```
+
+### Database
+
