--- api/app/routes/register.py
@@ -1,22 +1,0 @@
-from flask import request, make_response
-from datetime import datetime as dt
-from flask import current_app as app
-from ..models.user import User, db
-
-@app.route('/register', methods=['POST'])
-def user_records():
-    """Create a user via query string parameters."""
-    username = request.args.get('user')
-    email = request.args.get('email')
-    if username and email:
-        new_user = User(
-            username=username,
-            email=email,
-            created=dt.now(),
-            bio="In West Philadelphia born and raised, \
-            on the playground is where I spent most of my days",
-            admin=False
-        )
-        db.session.add(new_user)  # Adds new User record to database
-        db.session.commit()  # Commits all changes
-    return make_response(f"{new_user} successfully created!")
